# backend/tests/fixtures/openapi_before_mvc.json
{
  "components": {
    "schemas": {
...
      "CancelRideRequest": {
        "properties": {
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "CancelRideRequest",
        "type": "object"
      },
...
      "CreateRideRequest": {
        "properties": {
          "base_fare": {
            "title": "Base Fare",
            "type": "number"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "female_only": {
            "default": false,
            "title": "Female Only",
            "type": "boolean"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stops": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Stops",
            "type": "array"
          },
          "total_seats": {
            "default": 4,
            "title": "Total Seats",
            "type": "integer"
          }
        },
        "required": [
          "source",
          "destination",
          "base_fare"
        ],
        "title": "CreateRideRequest",
        "type": "object"
      },
      "EndRideRequest": {
        "properties": {
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          }
        },
        "title": "EndRideRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JoinRideRequest": {
        "properties": {
          "dropoff_stop": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dropoff Stop"
          },
          "pickup_stop": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickup Stop"
          },
          "seats": {
            "default": 1,
            "title": "Seats",
            "type": "integer"
          }
        },
        "title": "JoinRideRequest",
        "type": "object"
      },
...
      "UpdateStopStatusRequest": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "UpdateStopStatusRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Student-to-student ride-sharing network — Backend API",
    "title": "Arooohi API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/earnings/rides": {
      "get": {
        "description": "Per-ride earnings, newest first, with unsettled rides flagged.",
        "operationId": "earnings_rides_api_earnings_rides_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Earnings Rides",
        "tags": [
          "Driver Earnings"
        ]
      }
    },
...
    "/api/rides": {
      "get": {
        "description": "List rides: mine (as driver/passenger) + open available rides.",
        "operationId": "list_rides_api_rides_get",
        "parameters": [
          {
            "in": "query",
            "name": "female_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Female Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rides",
        "tags": [
          "Rides"
        ]
      },
      "post": {
        "description": "Driver creates a new ride with multi-stop and scheduling support.",
        "operationId": "create_ride_api_rides_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ride",
        "tags": [
          "Rides"
        ]
      }
    },
...
    "/api/rides/match": {
      "get": {
        "description": "Smart Matching Algorithm:\nMatches riders with open rides heading to their destination or passing through their intermediate stops.\nSupports proximity zone matching, class schedule time-flex matching, and female-only filtering.",
        "operationId": "match_rides_api_rides_match_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "pickup",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pickup"
            }
          },
          {
            "in": "query",
            "name": "destination",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Destination"
            }
          },
          {
            "in": "query",
            "name": "dropoff",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dropoff"
            }
          },
          {
            "in": "query",
            "name": "scheduled_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scheduled Date"
            }
          },
          {
            "in": "query",
            "name": "scheduled_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scheduled Time"
            }
          },
          {
            "in": "query",
            "name": "class_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class Time"
            }
          },
          {
            "in": "query",
            "name": "female_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Female Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Match Rides",
        "tags": [
          "Rides"
        ]
      }
    },
    "/api/rides/{ride_id}": {
      "get": {
        "description": "Ride detail including passenger stops and multi-stop progress.",
        "operationId": "get_ride_api_rides__ride_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ride",
        "tags": [
          "Rides"
        ]
      }
    },
    "/api/rides/{ride_id}/accept/{passenger_id}": {
      "post": {
        "description": "Driver accepts a passenger's ride request.",
        "operationId": "accept_passenger_api_rides__ride_id__accept__passenger_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "title": "Ride Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "passenger_id",
            "required": true,
            "schema": {
              "title": "Passenger Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Passenger",
        "tags": [
          "Rides"
        ]
      }
    },
    "/api/rides/{ride_id}/cancel": {
      "post": {
        "description": "Driver cancels the whole ride; a passenger cancels only their own seat.",
        "operationId": "cancel_ride_api_rides__ride_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Ride",
        "tags": [
          "Rides"
        ]
      }
    },
    "/api/rides/{ride_id}/cancellation-policy": {
      "get": {
        "description": "Preview what cancelling costs, so the UI can warn before anything happens.",
        "operationId": "cancellation_policy_api_rides__ride_id__cancellation_policy_get",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancellation Policy",
        "tags": [
          "Rides"
        ]
      }
    },
    "/api/rides/{ride_id}/end": {
      "post": {
        "description": "Driver ends the ride — computes distance (tracking points or estimate) for eco tracking.",
        "operationId": "end_ride_api_rides__ride_id__end_post",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndRideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "End Ride",
        "tags": [
          "Rides"
        ]
      }
    },
    "/api/rides/{ride_id}/join": {
      "post": {
        "description": "Rider requests a seat on a ride, optionally specifying their pickup and drop-off stops.",
        "operationId": "join_ride_api_rides__ride_id__join_post",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Join Ride",
        "tags": [
          "Rides"
        ]
      }
    },
    "/api/rides/{ride_id}/messages": {
      "get": {
        "description": "Ride Chat history (participants only).",
        "operationId": "get_ride_messages_api_rides__ride_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ride Messages",
        "tags": [
          "Rides"
        ]
      }
    },
    "/api/rides/{ride_id}/split": {
      "get": {
        "description": "Ride Cost Splitter — total = base_fare x surge; split by SEATS among accepted\npassengers, with whole-taka largest-remainder rounding so shares sum exactly\nto `total`.",
        "operationId": "ride_cost_split_api_rides__ride_id__split_get",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ride Cost Split",
        "tags": [
          "Rides"
        ]
      }
    },
    "/api/rides/{ride_id}/start": {
      "post": {
        "description": "Driver starts the ride.",
        "operationId": "start_ride_api_rides__ride_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "title": "Ride Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Ride",
        "tags": [
          "Rides"
        ]
      }
    },
    "/api/rides/{ride_id}/stops/{stop_id}/status": {
      "post": {
        "description": "Driver updates progress of a stop (pending, reached, departed).",
        "operationId": "update_stop_status_api_rides__ride_id__stops__stop_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "ride_id",
            "required": true,
            "schema": {
              "title": "Ride Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "title": "Stop Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStopStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Stop Status",
        "tags": [
          "Rides"
        ]
      }
    },
...
  }
}